tests/wpt/tests/html/semantics/embedded-content/the-frame-element/document-getters-return-null-for-cross-origin.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test that contentDocument returns null for a cross-origin document.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test('HTMLFrameElement.contentDocument for cross-origin document');
window.addEventListener(
    'load', t.step_func_done(() => { assert_equals(document.querySelector('frame').contentDocument, null); }));
</script>
<frameset>
<frame src='data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"><rect height="100" width="100"/></svg>'></frame>
</frameset>
